Update LICENSE
Added custom MIT license defining what software means (guides not code) in my repo and making it mandatory that: My guides remain unedited, credit be given anywhere they are used and if the individual wishes to use them or especially wishes to alter them for redistribution they must provide a link to my repository where the original guides can be found and credit to me. No credit markers at the beginning of guides can be removed if you wish to modify and redistribute this project.